+++ infrastructure/terraform/per_account/dev/vars.tf
@@ -28,3 +28,18 @@ variable "budget_limit" {
   default = "1050"
   type    = string
 }
+
+# BACKUPS_LOGIC
+variable "python_version" {
+  default = "python3.12"
+}
+
+# BACKUPS_LOGIC
+variable "layers" {
+  type = list(string)
+}
+
+# BACKUPS_LOGIC
+variable "third_party_layers" {
+  type = list(string)
+}